Patch from Isak Styf (SourceForge user istyf) for Automatic MacOSX locale. Isak says:
Since there is a user setting in the MacOS X system preferences to select the preferred language in applications. Why not let BZFlag take that setting into consideration when displaying the default locale to the user?

The loadBundle part of BundleMgr (ifdef _MACOSX_) checks if the wanted locale is "default". If so, it uses the CoreFoundation APIs to locate the po-file that belongs to the selected locale and overwrites the path that is generated in all other cases. As usual if there are no files for any of the languages in the users priority list (yes, you can select many) BZFlag reverts to english.

(The default locale can be overridden as before by selecting a different locale in the in-game GUI options, but as long as the locale is set to "default", the game will use the language specified by the user in the system prefs.)

The locale is initially set to "default" if none is specified in the .bzflag config file.

Apart from theese two small changes I have also included a fix to remove cstdio from the MacOSX pch-file and a fixup of a possible (?) typo where a comment before #define ROBOTS have no enclosing */ thus causing the ROBOT code not to be compiled in.

The ProjectBuilder file has had an invalid executable reference removed as well as a bunch of CopyFiles phases added to copy the language files into their respective .lproj subfolders as the CF API demands. Note that the en.lproj subfolder is created with nothing but a placeholding rgb file inside of it, this is because if the en.lproj directory doesnt exist, BZFlag will use the next available language from the users list of languages. So dont delete that buildphase even if it seems braindead.


git-svn-id: https://svn.code.sf.net/p/bzflag/code/branches/v1_7@1274 08b3d480-bf2c-0410-a26f-811ee3361c24