rewrite the keyboard angular velocity dampening to take into consideration the current framerate.  this allows for consistent dampening regardless of client framerate (whereas before, high framerates become practically instant and low framerates unbearable).  we ditch the prior method that converged increasingly slower towards the desired rate and now actually accelerate towards the desired rate within some configurable timeframe.
currently set to a window of roughly .5 to .25 seconds, the dampening will converge to the desired rate within about .35 seconds depending on the starting angualar velocity.  there is debug code tied to the bzdb debugNewAngVel setting that will toggle the display of informative values that tick to bz console and terminal shell.

this approach isn't as exact as the direction randomparticle was going, but it should be more than sufficient (and a faster approximation).  this change was inspired by randomparticle's work though this implementation was separate.  thx for the kick-off, comments and additional testing welcome.


git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk@13708 08b3d480-bf2c-0410-a26f-811ee3361c24