Fix stack corruption when plugin load fails
fastmap depends on the HTTPServer plugin.  If the HTTPServer plugin is
not available the BZFSHTTP() constructor is called repeatedly which
blows the stack and causes a segfault.

If we can't find the plugin don't try to load it - just issue a sane
error message instead.

On Linux:

bzfs -loadplugin plugins/.libs/fastmap

fails because HTTPServer is not in the plugin path.  The resulting
segfault isn't much fun to deal with.

After this patch bzfs reports

HTTP Plugin fastmap failed to load due to callback failure, make sure the HTTPServer plugin is in the same path

which is correct.

git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk@20678 08b3d480-bf2c-0410-a26f-811ee3361c24